feat: Second modal has been replaced
