+++ src/components/BurgerConstructor/BurgerConstructor.tsx
@@ -9,8 +9,7 @@ import { FC, useMemo, useState } from "react";
 import { IBurgerConstructorProps } from "./types.ts";
 import { Modal } from "../Modal";
 import { OrderDetails } from "./components/OrderDetails";
-import { IngredientDetails } from "@components/BurgerConstructor/components/IngredientDetails";
-import { IIngredient } from "@api/getIngredients";
+
 export const BurgerConstructor: FC<IBurgerConstructorProps> = ({
   selected,
   setSelected,
@@ -19,17 +18,6 @@ export const BurgerConstructor: FC<IBurgerConstructorProps> = ({
     return selected.reduce((acc, item) => acc + item.price, 0);
   }, [selected]);
   const [isOpen, setOpen] = useState(false);
-  const [selectedItem, setSelectedItem] = useState<IIngredient>(
-    {} as IIngredient,
-  );
-
-  const handleOpenModal = (item: IIngredient) => {
-    setSelectedItem(item);
-  };
-
-  const handleCloseIngredientDetails = () => {
-    setSelectedItem({} as IIngredient);
-  };
 
   const handleDelete = (idx: number) => {
     const newState = structuredClone(selected);
@@ -54,23 +42,17 @@ export const BurgerConstructor: FC<IBurgerConstructorProps> = ({
           const listType =
             type === "bun" ? (idx === 0 ? "top" : "bottom") : undefined;
           return (
-            <div
-              onClick={() => {
-                handleOpenModal(item);
+            <ConstructorElement
+              key={name + idx}
+              text={name}
+              thumbnail={image_mobile}
+              price={price}
+              isLocked={isLocked}
+              type={listType}
+              handleClose={() => {
+                handleDelete(idx);
               }}
-            >
-              <ConstructorElement
-                key={name + idx}
-                text={name}
-                thumbnail={image_mobile}
-                price={price}
-                isLocked={isLocked}
-                type={listType}
-                handleClose={() => {
-                  handleDelete(idx);
-                }}
-              />
-            </div>
+            />
           );
         })}
       </div>
@@ -85,12 +67,6 @@ export const BurgerConstructor: FC<IBurgerConstructorProps> = ({
       )}
       <Modal isOpen={isOpen} onClose={handleCloseOrderDetails}>
         <OrderDetails />
-      </Modal>
-      <Modal
-        isOpen={!!Object.keys(selectedItem).length}
-        onClose={handleCloseIngredientDetails}
-      >
-        <IngredientDetails item={selectedItem} />
       </Modal>
     </section>
   );
